admin 80%

+++ collection.php
@@ -5,6 +5,14 @@
         header("Location: logreg.php");
         exit;
     }
+
+    if(!empty($_GET['condition'])){
+        $_SESSION['undisplay'] = True;
+    }
+    
+    else {
+        $_SESSION['undisplay'] = False;
+    }
 ?>
 
 <!DOCTYPE html>
@@ -70,7 +78,11 @@
         <div class="col mx-2">
             <nav class="navbar navbar-light bg-light justify-content-between">
                 <div class="form-inline justify-content-start">
-                    <button id="filter_btn" class="btn btn-dark mr-sm-2" role="button" onclick="showFilter()">Filter</button>
+                    <?php
+                        if($_SESSION['level'] == 2){
+                            echo '<button id="filter_btn" class="btn btn-dark mr-sm-2" role="button" onclick="showFilter()">Filter</button>';
+                        }
+                    ?>
                     <form class="form-inline">
                         <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
                         <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
@@ -78,52 +90,57 @@
                 </div>
 
                 <?php
-                    $user_id = $_SESSION['id'];
+                    if ($_SESSION['level'] == 2){
 
-                    if (isset($_POST['col_mode'])){
-                        $query = mysqli_query($con, "SELECT MAX(id) FROM borrow_book");
-                        $borrow_id = mysqli_fetch_row($query);
-                        mysqli_query($con, "DELETE FROM borrow_book WHERE id='$borrow_id[0]' AND user_id='$user_id'");
-                        $_SESSION['borrow'] = NULL;
+                        $user_id = $_SESSION['id'];
+
+                        if (isset($_POST['col_mode'])){
+                            $query = mysqli_query($con, "SELECT MAX(id) FROM borrow_book");
+                            $borrow_id = mysqli_fetch_row($query);
+                            mysqli_query($con, "DELETE FROM borrow_book WHERE id='$borrow_id[0]' AND user_id='$user_id'");
+                            $_SESSION['borrow'] = NULL;
+                        }
+
+                        if (isset($_POST['bor_mode'])){
+                            mysqli_query($con, "INSERT INTO borrow_book (user_id) VALUES ('$user_id')");
+                            $_SESSION['borrow'] = True;
+                        }
+
+                        if(isset($_SESSION['borrow'])){
+                            echo '
+                            <div class="form-inline justify-content-end">
+                                <button id="cart_btn" class="btn btn-primary mr-sm-2" role="button" onclick="showCart()">Check Cart</button>
+                                <form class="form-inline" method="POST" action="">
+                                    <input class="btn btn-danger form-control mr-sm-2" type="submit" name="col_mode" value="Cancel">
+                                </form>
+                            </div>';
+                        }
+
+                        else{
+                            echo '
+                            <div class="form-inline justify-content-end">
+                                <form class="form-inline" method="POST" action="">
+                                    <input class="btn btn-success form-control mr-sm-2" type="submit" name="bor_mode" value="Pinjam Buku">
+                                </form>
+                                <form class="form-inline" method="POST" action="historyCollection.php">
+                                    <input class="btn btn-info form-control mr-sm-2" type="submit" name="history" value="Riwayat Peminjaman">
+                                </form>
+                            </div>';
+                        }
                     }
-
-                    if (isset($_POST['bor_mode'])){
-                        mysqli_query($con, "INSERT INTO borrow_book (user_id) VALUES ('$user_id')");
-                        $_SESSION['borrow'] = True;
-                    }
-
-                    if(isset($_SESSION['borrow'])){
-                        echo '
-                        <div class="form-inline justify-content-end">
-                            <button id="cart_btn" class="btn btn-primary mr-sm-2" role="button" onclick="showCart()">Check Cart</button>
-                            <form class="form-inline" method="POST" action="">
-                                <input class="btn btn-danger form-control mr-sm-2" type="submit" name="col_mode" value="Cancel">
-                            </form>
-                        </div>';
-                    }
-
-                    else{
-                        echo '
-                        <div class="form-inline justify-content-end">
-                            <form class="form-inline" method="POST" action="">
-                                <input class="btn btn-success form-control mr-sm-2" type="submit" name="bor_mode" value="Pinjam Buku">
-                            </form>
-                            <form class="form-inline" method="POST" action="historyCollection.php">
-                                <input class="btn btn-info form-control mr-sm-2" type="submit" name="history" value="Riwayat Peminjaman">
-                            </form>
-                        </div>';
-                    }
-
                 ?>
                 
             </nav>
 
-            <div class="row">
-                <?php 
+            <div class="row mt-3">
+                <?php
+                    if ($_SESSION['level'] == 1){
+                        include "adminBar.php";
+                    }
+                    
                     include "readCollection.php";
                 ?>
             </div>
-            
         </div>
 
         <div id="cart_block" class="col-2">
